requirements_val.txt
streamlit>=1.30,<2
pandas>=2.0
python-dotenv>=1.0
google-generativeai>=0.8.3
google-genai>=0.3.0
google-auth>=2.29
pymilvus>=2.4.0
audio-recorder-streamlit
pydub
Pillow
protobuf>=4.25,<6
grpcio>=1.62